feat: Finish logout implementation
Success responses from login and logout was changed to json
because they are processed by JS scripts in the frontend.
This way, the double redirect problem is solved.

Code was cleaned given that experimentation took place while
trying to solve this issue. Login button was properly
implemented and unnecessary gapi calls were removed.

Resolves: #29